java functional programs

+++ src/com/bridgelabz/javaconditionprograms/Sin.java
@@ -9,7 +9,7 @@ public class Sin {
 	public static void main(String[] args) {
 		Scanner scanner=new Scanner(System.in);
 		System.out.println("Enter the angle: ");
-		double angle=scanner.nextInt();
+		double angle=scanner.nextDouble();
 		Utility.sin(angle);
 		   scanner.close();     
 	    }
